+++ CHANGELOG.md
@@ -15,7 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
-- Use `testing.TB` interface instead of `*testing.T` struct in metricdatatest. (#4483)
+- Use a `TB` interface instead of `*testing.T` struct in metricdatatest. (#4483)
 
 ### Deprecated